# lib/api-zod/src/generated/types/tiktokInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TikTok HD Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface TiktokInput {
  /** TikTok video URL */
  url: string;
}