# lib/api-zod/src/generated/types/tiktokVideo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TikTok HD Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface TiktokVideo {
  id: string;
  title: string;
  author: string;
  authorAvatar: string;
  cover: string;
  /** Watermark-free HD download URL */
  downloadUrl: string;
  /** Duration in seconds */
  duration: number;
  likes: number;
  comments: number;
  shares: number;
}